fix: clippy + test serialization for YT health (Refs #194)
Clippy fixes:
- delivery_status.rs: collapse `if status == 403/429` match guards into
  literal pattern matches (clippy::redundant_guards).
- youtube.rs / delivery_youtube.rs / streams.rs: drop redundant
  `secs as i64` casts now that TokenResponse.expires_in is i64
  (clippy::unnecessary_cast).

Test serialization:
- Add shared yt_health_test_env::env_guard tokio::Mutex so parallel
  #[tokio::test] instances do not race on the process-global
  YOUTUBE_API_BASE env var (one test's reqwest call would otherwise
  hit another test's wiremock URL).
- Apply the gate in delivery_status_yt_health_tests, yt_health_cache_tests,
  and (a separate per-crate gate) streams_for_label_tests.
- Add `clear_yt_health_cache_for_test` so the 15s in-process cache does
  not leak between tests that allocate endpoint id=1 in fresh pools.

+++ crates/rs-api/src/delivery_youtube.rs
@@ -45,7 +45,7 @@ impl DeliveryOrchestrator {
             match oauth::refresh_access_token(&oauth_tokens).await {
                 Ok(resp) => {
                     let new_expires = resp.expires_in.map(|secs| {
-                        (chrono::Utc::now() + chrono::Duration::seconds(secs as i64)).to_rfc3339()
+                        (chrono::Utc::now() + chrono::Duration::seconds(secs)).to_rfc3339()
                     });
 
                     if let Err(e) = db::upsert_youtube_oauth(
